pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kaustubh <kaustubhkagrawal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.2"
openai = "^1.11.1"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"